Refactor Code & Minor Bug fixes
