--- commonHandler.js
@@ -1,4 +1,0 @@
-export const routeNotFound = (res) => {
-    res.statusCode = 404;
-    res.end(JSON.stringify({ error: 'Not Found' }))
-}
