+++ globalFunctions.js
@@ -1,3 +1,8 @@
+export const routeNotFound = (res) => {
+    res.statusCode = 404;
+    res.end(JSON.stringify({ error: 'Not Found' }))
+}
+
 export const generateId = (t) => {
     const timestamp = Date.now();
